test: add excelToDate utility tests
